+++ app/controllers/topics_controller.rb
@@ -18,10 +18,16 @@ class TopicsController < ApplicationController
   end
 
   def create
-    @topic = Topic.new
-    @topic.name = params[:topic][:name]
-    @topic.description = params[:topic][:description]
-    @topic.public = params[:topic][:public]
+
+    @topic = Topic.new(topic_params)
+
+
+    if current_user.moderator?
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to action: :index
+      return
+    end
+
 
     if @topic.save
        redirect_to @topic, notice: "Topic was saved successfully."
@@ -38,9 +44,8 @@ class TopicsController < ApplicationController
    def update
     @topic = Topic.find(params[:id])
 
-    @topic.name = params[:topic][:name]
-    @topic.description = params[:topic][:description]
-    @topic.public = params[:topic][:public]
+    @topic.assign_attributes(topic_params)
+
 
     if @topic.save
        flash[:notice] = "Topic was updated."
@@ -53,6 +58,12 @@ class TopicsController < ApplicationController
 
   def destroy
     @topic = Topic.find(params[:id])
+
+    if current_user.moderator?
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to action: :index
+      return
+    end
 
     if @topic.destroy
       flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
@@ -70,7 +81,7 @@ class TopicsController < ApplicationController
   end
 
   def authorize_user
-    unless current_user.admin?
+    unless current_user.admin? || current_user.moderator?
       flash[:alert] = "You must be an admin to do that."
       redirect_to topics_path
     end
